requirements.txt
chainlit
openai
requests
python-dotenv
